Add Caribou to GTK_MODULES
The Caribou IM module would override any other modules the user is
utilizing; by adding it to the GTK_MODULES the user can decide whether
or not to enable the module (enabled automatically for onscreen
keyboard) and will avoid override. There is also no dependency for
the abstract class GtkIMContextSimple.